Added error tracking into the DHCPResponse data structure
Added error tracking into the DHCPResponse data structure.  Then restructured main() to use this as the location to store information on mismatches.

I did make some adjustments to the on screen output.  The data point includes the '<-- BAD' tag if it is a problematic field as per tradition.  But I moved the expected values each to their own line, (one per server that doens't match), prepended with the traditional '!!!'
This should resolve "When using multiple servers a mis-match poisons the well for future checks. #7" so I'm bumping the version number to 1.2